fix bug

+++ database/migrations/2025_10_20_102625_add_foto_profil_to_users_table.php
@@ -6,20 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-public function up(): void
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->string('foto_profil')->nullable()->after('email');
-    });
-}
+    public function up(): void
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users', 'foto_profil')) {
+                $table->string('foto_profil')->nullable()->after('email');
+            }
+        });
+    }
 
-public function down(): void
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->dropColumn('foto_profil');
-    });
-}
+    public function down(): void
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasColumn('users', 'foto_profil')) {
+                $table->dropColumn('foto_profil');
+            }
+        });
+    }
 };
